use size_t for lengths

+++ src/core/stem.h
@@ -1,6 +1,8 @@
 #ifndef CORE_STEM_H
 #define CORE_STEM_H
 
-int stem (char *);
+#include <stdlib.h>
+
+size_t stem (char *);
 
 #endif
